Renamed test package to e2eTests as these are e2e tests, not normal junit tests.
