+++ sdk-tests/src/test/java/nl/tytech/sdk/e2eTests/MyStakeholder.java
@@ -1,18 +1,10 @@
-package nl.tytech.sdk.example;
+package nl.tytech.sdk.e2eTests;
 
 import static org.junit.Assert.assertNotNull;
 import static org.junit.Assert.assertTrue;
 
-import java.util.HashMap;
-import java.util.Map;
-
-import javax.security.auth.login.LoginException;
-
 import com.vividsolutions.jts.geom.MultiPolygon;
 
-import login.Login;
-import login.ProjectException;
-import login.ProjectFactory;
 import nl.tytech.core.client.event.EventManager;
 import nl.tytech.core.client.net.ServicesManager;
 import nl.tytech.core.client.net.TSlotConnection;
@@ -35,53 +27,12 @@ import nl.tytech.data.engine.item.Stakeholder.Type;
 import nl.tytech.locale.TLanguage;
 
 /**
- * Game field where general testing of interactions can be done for testing.
- * 
- * @author W.Pasman
- *
- */
-public class GameField {
-
-	private ProjectData project;
-	private Map<Stakeholder.Type, MyStakeholder> stakeholders = new HashMap<Stakeholder.Type, MyStakeholder>();
-
-	public GameField() throws LoginException, ProjectException {
-		Login login = new Login();
-		login.doLogin();
-
-		ProjectFactory factory = new ProjectFactory();
-		String projectName = "test" + System.currentTimeMillis();
-		project = factory.createProject(projectName);
-		if (project == null) {
-			throw new IllegalStateException("createProject returned null");
-		}
-	}
-
-	/**
-	 * Call this once for each expected stakeholder. You may then use the
-	 * returned stakeholder immediately
-	 * 
-	 */
-	public MyStakeholder addStakeholder(Stakeholder.Type type) {
-		MyStakeholder stakeholder = new MyStakeholder(type, project);
-		stakeholders.put(type, stakeholder);
-		return stakeholder;
-	}
-
-	public void close() throws ProjectException {
-		new ProjectFactory().deleteProject(project);
-		project = null;
-	}
-
-}
-
-/**
  * Test stakeholder.
  * 
  * @author W.Pasman
  *
  */
-class MyStakeholder {
+public class MyStakeholder {
 
 	private Type type;
 	private ExampleEventHandler events;
