+++ sdk-tests/src/test/java/nl/tytech/sdk/e2eTests/PlanRoadsAndBuildingsTest.java
@@ -1,4 +1,4 @@
-package nl.tytech.sdk.example;
+package nl.tytech.sdk.e2eTests;
 
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertNotNull;
